+++ tasks/w-cmd.js
@@ -3,6 +3,7 @@ var color = require('yyl-color');
 var util = require('./w-util.js');
 var vars = util.vars;
 
+
 var
   events = {
     version: require('./w-version'),