+++ tasks/w-test.js
@@ -1,8 +1,5 @@
 'use strict';
-const gulpfile = require('../init-files/gulp-requirejs/index.js');
-
 const wTest = function() {
-  gulpfile.run();
 };
 
 module.exports = wTest;