+++ tasks/w-vars.js
@@ -1,8 +1,0 @@
-'use strict';
-var util = require('./w-util.js');
-var path = require('path');
-
-module.exports = util.extend(util.vars, {
-  BASE_PATH: path.join(__dirname, '..')
-});
-